Shrink structure size

+++ game/game/include/game/unit_defs.h
@@ -3,7 +3,7 @@
 #include "unit_types.h"
 
 struct UnitDef {
-  u16 movement;
+  u8 movement;
 };
 
 extern const UnitDef unitDefs[static_cast<u16>(UnitType::count)];
